% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy.a-package.R
\docType{package}
\name{pregnancy.a-package}
\alias{pregnancy.a}
\alias{pregnancy.a-package}
\title{pregnancy.a: Research Compendium of the GIPSO's Sleep and Pregnancy SQLR}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R package with the research compendium of the GIPSO's sleep and pregnancy project: a systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/pregnancy.a/}
  \item \url{https://github.com/gipso/pregnancy.a/}
  \item Report bugs at \url{https://github.com/gipso/pregnancy.a/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alicia Rafaelly Vilefort Sales \email{alicia.sales@usp.br} (\href{https://orcid.org/0000-0003-0004-4487}{ORCID}) [copyright holder]
  \item Isadora Rodrigues \email{isadorard@usp.br} (\href{https://orcid.org/0000-0002-3629-6208}{ORCID})
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
