% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.pregnancy.a-package.R
\docType{package}
\name{rc.pregnancy.a-package}
\alias{rc.pregnancy.a}
\alias{rc.pregnancy.a-package}
\title{rc.pregnancy.a: Research Compendium of the GIPSO Sleep and Pregnancy SQLR}
\description{
Research compendium of the Interdisciplinary Sleep Research Group (GIPSO) sleep and pregnancy systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/rc.pregnancy.a/}
  \item \url{https://github.com/gipso/rc.pregnancy.a/}
  \item Report bugs at \url{https://github.com/gipso/rc.pregnancy.a/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alicia Rafaelly Vilefort Sales \email{alicia.sales@usp.br} (\href{https://orcid.org/0000-0003-0004-4487}{ORCID}) [copyright holder]
  \item Isadora Rodrigues \email{isadorard@usp.br} (\href{https://orcid.org/0000-0002-3629-6208}{ORCID})
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
